wip(input): partial Switch Pro Bluetooth report handling
Splits the Switch Pro report parsers into switch-pro-hid.ts, matching how
sony-hid.ts already separates the DualShock and DualSense layouts, and
adds a branch for the padded ~362 byte report a Bluetooth-connected pad
sends under report id 0x3f.

Incomplete on purpose. Buttons are shifted one byte from the USB layout
and are believed correct. Axes and the d-pad are stubbed to neutral: the
byte offsets guessed from a single capture produced a stick reading full
deflection at rest, which is worse than reporting nothing.

SDL's own struct for this report says the axes are four 16 bit little
endian values centred on 0x8000 and the d-pad is a standard hat, so the
correct layout is known; it is not applied here because it has not been
tested against hardware. See plans/sdl3-input-binding.html for the wider
question of whether this file should exist at all.
